+++ appinfo/routes.php
@@ -13,6 +13,7 @@ return [
         ['name' => 'activity#destroy',     'url' => '/api/activities/{id}',             'verb' => 'DELETE'],
         ['name' => 'activity#trackpoints', 'url' => '/api/activities/{id}/trackpoints', 'verb' => 'GET'],
         ['name' => 'activity#laps',        'url' => '/api/activities/{id}/laps',        'verb' => 'GET'],
+        ['name' => 'activity#photos', 'url' => '/api/activities/{id}/photos', 'verb' => 'GET'],
 
         // Sleep API
         ['name' => 'sleep#index',   'url' => '/api/sleep',             'verb' => 'GET'],